backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
celery[redis]==5.4.0
redis==5.1.0
langchain==0.3.0
langchain-huggingface>=1.0.0
langchain-community==0.3.0
langchain-chroma>=0.1.4
langchain-text-splitters==0.3.0
langchain-core>=0.3.0
chromadb>=1.3.5
huggingface-hub>=0.24.0
pypdf==4.3.0
psutil==6.0.0
pydantic==2.9.0
pydantic-settings==2.5.0
websockets==13.0
python-dotenv>=1.0.0
